[Feature] : 위치 추가 기능 구현 완료.

+++ IveBeenHere/entity/PlaceDTO.swift
@@ -16,7 +16,7 @@ struct SnapInfo: Codable {
 }
 
 struct PlaceDTO: Codable {
-    let latitude: Int
+    let latitude: Double
     let name: String
-    let longitude: Int
+    let longitude: Double
 }
